// max-mini-app/src/api/models/Queue.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MaxQueue Orchestrator API
 * API, which is the entry point to the MaxQueue project
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Queue
 */
export interface Queue {
    /**
     * 
     * @type {string}
     * @memberof Queue
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof Queue
     */
    name?: string;
}

/**
 * Check if a given object implements the Queue interface.
 */
export function instanceOfQueue(value: object): value is Queue {
    return true;
}

export function QueueFromJSON(json: any): Queue {
    return QueueFromJSONTyped(json, false);
}

export function QueueFromJSONTyped(json: any, ignoreDiscriminator: boolean): Queue {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function QueueToJSON(json: any): Queue {
    return QueueToJSONTyped(json, false);
}

export function QueueToJSONTyped(value?: Queue | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
    };
}